agents/a2a_agents/samples/travel_agent/pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "langchain-google-genai>=2.1.12",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
